Move isotp out of libaktualizr
