+++ src/CMakeLists.txt
@@ -8,6 +8,7 @@ add_subdirectory("virtual_secondary")
 add_subdirectory("sota_tools")
 add_subdirectory("aktualizr_primary")
 add_subdirectory("libaktualizr-posix")
+add_subdirectory("libaktualizr-isotp")
 add_subdirectory("aktualizr_secondary")
 add_subdirectory("aktualizr_info")
 add_subdirectory("uptane_generator")
