+++ src/libaktualizr-c/CMakeLists.txt
@@ -16,9 +16,6 @@ add_library(aktualizr-c SHARED ${SOURCES}
     $<TARGET_OBJECTS:logging>
     $<TARGET_OBJECTS:storage>
     $<TARGET_OBJECTS:uptane>)
-if (BUILD_ISOTP)
-    target_sources(aktualizr-c PRIVATE $<TARGET_OBJECTS:isotp_conn>)
-endif (BUILD_ISOTP)
 
 target_include_directories(aktualizr-c PUBLIC ${PROJECT_SOURCE_DIR}/include)
 target_link_libraries(aktualizr-c PRIVATE ${AKTUALIZR_EXTERNAL_LIBS})
