+++ src/libaktualizr/CMakeLists.txt
@@ -14,10 +14,6 @@ add_subdirectory("storage")
 add_subdirectory("package_manager")
 add_subdirectory("telemetry")
 
-if(BUILD_ISOTP)
-    add_subdirectory("isotp_conn")
-endif(BUILD_ISOTP)
-
 # deprecated, we recommend using aktualizr_lib
 add_library(aktualizr_static_lib STATIC
     $<TARGET_OBJECTS:asn1>
@@ -55,8 +51,3 @@ add_library(aktualizr_lib SHARED
 target_link_libraries(aktualizr_lib ${AKTUALIZR_EXTERNAL_LIBS})
 set_target_properties(aktualizr_lib PROPERTIES LIBRARY_OUTPUT_NAME aktualizr)
 install(TARGETS aktualizr_lib LIBRARY DESTINATION lib COMPONENT aktualizr)
-
-if (BUILD_ISOTP)
-    target_sources(aktualizr_static_lib PRIVATE $<TARGET_OBJECTS:isotp_conn>)
-    target_sources(aktualizr_lib PRIVATE $<TARGET_OBJECTS:isotp_conn>)
-endif (BUILD_ISOTP)
